bench/python-sqlite3/requirements.txt
python-rapidjson == 1.5
